requirements.txt
aiohttp==3.7.4.post0
akinator
asyncio==3.4.3
beautifulsoup4
bs4==0.0.1
discord.py==2.1.0
giphy-client==1.0.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-search-results==2.4.1
googleapis-common-protos==1.56.1
googlesearch-python==1.1.0
googletrans==4.0.0rc1
Janome==0.4.2
numpy==1.22.3
orjson==3.6.7
PyNaCl==1.4.0
requests==2.25.1
requests-html==0.10.0
selenium==4.1.5
sympy==1.10
typing==3.7.4.3
ujson==5.4.0
wavelink==1.3.1
wikipedia==1.4.0
simpcalci==0.0.2
jishaku==2.5.1
psycopg2==2.9.3
Pillow==9.1.1
pyzbar==0.1.9
qrcode==7.3.1
